backend/payment_risk_shield_be/requirements.txt
Flask==3.0.3
python-whois
pandas
requests==2.28.1
beautifulsoup4==4.11.1
python-whois==0.7.3
joblib
xgboost